renamed createAssessment into createAssessmentFile
